The Rework, pt. 8
Removed the duplicate implementation of gson

Added more config options

Reenabled initTypes, which has been entirely reworked. Only triggers if a new config option is true.

IOInit now registers metals directly to MFR's metal registry, bypassing any writing/reading.

Scribe now stores the metals written to it as an array.

Rewrote all of the mfrAPI classes to better match MFR

Added several factories, beginning implementation of JSON togglability

Changes existing json recipes to stop using direct mentions of Embers ores, to better support their removal. Added factories to toggle Electrum alloy crucible recipe and spark plug--they need fixing still

+++ src/main/java/com/Lenvill/mfrAPI/MaterialArmorStats.java
@@ -1,34 +1,34 @@
 package com.Lenvill.mfrAPI;
 
 public class MaterialArmorStats {
-    double cutting;
-    double blunt;
-    double piercing;
+    float cutting;
+    float blunt;
+    float piercing;
 
     //Getters
-    public double getCutting() {
+    public float getCutting() {
         return cutting;
     }
 
-    public double getBlunt() {
+    public float getBlunt() {
         return blunt;
     }
 
-    public double getPiercing() {
+    public float getPiercing() {
         return piercing;
     }
 
 
     //Setters
-    public void setCutting(double cutting) {
+    public void setCutting(float cutting) {
         this.cutting = cutting;
     }
 
-    public void setBlunt(double blunt) {
+    public void setBlunt(float blunt) {
         this.blunt = blunt;
     }
 
-    public void setPiercing(double piercing) {
+    public void setPiercing(float piercing) {
         this.piercing = piercing;
     }
 }
